refactor: 알림 설정 클래스 분리
